requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
cachetools==5.5.0
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
ollama==0.3.3
